pyproject.toml
[project]
name = "lc-calendar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "plotly>=6.5.0",
    "python-dateutil>=2.9.0.post0",
]

[project.scripts]
lccal = "lc_calendar.cli:main"
